Update: fetch spotify now playing song
