+++ package.json
@@ -20,7 +20,6 @@
     "contentful": "^9.1.28",
     "framer-motion": "^4.1.17",
     "next": "^11.1.2",
-    "querystring": "^0.2.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-icons": "^4.2.0",
